src/main/java/models/Location.java
package models;

import java.util.Objects;

public class Location {

    private String name;
    private int id;

    public Location(String name){
        this.name = name;
    }

    public String getName() { return name; }

    public int getId() { return id; }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Location location = (Location) o;
        return getId() == location.getId() &&
                Objects.equals(getName(), location.getName());
    }

    @Override
    public int hashCode() {
        return Objects.hash(getName(), getId());
    }
}
